+++ src/GetSitesWithProfile.c
@@ -63,7 +63,7 @@ long  GetSitesWithProfile(char* s,
 			  index = OligoToInt(s+j, p->order+1,5);
 			  
 			  if (index >= p->dimensionTrans)
-				score = score + -INFI;
+				score = score + -(float)INFI;
 			  else
 				score = score + p->transitionValues[i][index];
 			}
@@ -100,7 +100,7 @@ long  GetSitesWithProfile(char* s,
 			  /* i is the position inside the region */
 			  index = TRANS[(int)(*(s + i))];
 			  if (index >= p->dimensionTrans)
-				score = score + -INFI;
+				score = score + -(float)INFI;
 			  else
 				score = score + p->transitionValues[i][index];
 			}
@@ -132,7 +132,7 @@ long  GetSitesWithProfile(char* s,
 			  /* i is the position inside the region */
 			  index = 5*TRANS[(int)(*(s + i -1))] + TRANS[(int)(*(s + i))];
 			  if (index >= p->dimensionTrans)
-				score = score + -INFI;
+				score = score + -(float)INFI;
 			  else
 				score = score + p->transitionValues[i][index];
 			}
@@ -166,7 +166,7 @@ long  GetSitesWithProfile(char* s,
 			  index = OligoToInt(s + i - p->order, p->order+1,5);
 
 			  if (index >= p->dimensionTrans)
-				score = score + -INFI;
+				score = score + -(float)INFI;
 			  else
 				score = score + p->transitionValues[i][index];
 			}
